Cleanup SPIFloat and default values in styles
Remove remaining uses of SCALE24_TO_FLOAT and FROM_FLOAT macros which were only
used for SPIFloat and can easily be folded into the object itself.

Harmonize the default data type use with SPIEnum, SPILength and SPIFontSize
this allows the rendering engine nr-style.cpp to use the right rendering value
which is always the computed value and not need to think about the internal
implementation of the style variable between `value`, `computed` and will allow
these variable to be made private in the future.

Mikekov: Add `as_double` and `as_enum` getter functions for the new explicit interface.

+++ src/display/nr-style.h
@@ -69,7 +69,7 @@ struct NRStyleData
     Paint stroke;
     float stroke_width;
     bool hairline;
-    float miter_limit;
+    double miter_limit;
     std::vector<double> dash;
     float dash_offset;
     cairo_fill_rule_t fill_rule;
